Extract snapshot utils into module
